FIX: SlideBanner 리팩토링
