+++ src/libs/Preload/index.tsx
@@ -1,0 +1,6 @@
+export function preloading(imageArray:string[]) {
+    imageArray.forEach((url) => {
+      const image = new Image();
+      image.src = url;
+    });
+  }
