+++ src/libs/SlideHook/index.tsx
@@ -1,0 +1,56 @@
+import { useEffect, useRef } from "react";
+
+interface IUseInterval {
+  (callback: () => void, interval: number): void;
+}
+
+// box 애니매이션 variant
+export const variants = {
+  enter: (direction: number) => {
+    return {
+      x: direction > 0 ? 200 : -200,
+      opacity: 0,
+    };
+  },
+  center: {
+    zIndex: 1,
+    x: 0,
+    opacity: 0.55,
+  },
+  exit: (direction: number) => {
+    return {
+      zIndex: 0,
+      x: direction < 0 ? 200 : -200,
+      opacity: 0,
+    };
+  },
+};
+
+// setInterval을 사용해 상태 변화를 만들 수 있는 hook
+export const useInterval: IUseInterval = (callback, interval) => {
+  const savedCallback = useRef<(() => void) | null>(null);
+
+  useEffect(() => {
+    savedCallback.current = callback;
+  });
+
+  useEffect(() => {
+    function tick() {
+      if (savedCallback.current) {
+        savedCallback.current();
+      }
+    }
+
+    // interval이 10초라는 예외의 숫자가 되었을 때 setInterval을 동작하지 않도록 함
+    if (interval !== 10000) {
+      let id = setInterval(tick, interval);
+      return () => clearInterval(id);
+    }
+  }, [interval]);
+};
+
+// drag 슬라이드 용 함수와 변수
+export const swipeConfidenceThreshold = 10000;
+export const swipePower = (offset: number, velocity: number) => {
+  return Math.abs(offset) * velocity;
+};
